changed regular expression
The code didn't work for me (bash v4.2.25). I struggled for a bit learning bash regexp, then I came out with that.

So now the expression is : "access_token":"<any number of char different of ">", which seems cleaner and will possibly run in more shells.

Also, I used a variable for the regex, which is recommended and allows to forget the tedious double quotes escaping.